\name{bayesTFR-internal}
\alias{burn.and.thin}
\alias{bdem.map.gvis}
\alias{bdem.parameter.traces}
\alias{diag.thin.indep} 
\alias{DLcurve}
\alias{do.convert.trajectories}
\alias{do.get.tfr.parameter.traces}
\alias{do.get.traces}
\alias{do.get.traces.bayesTFR.mcmc}
\alias{do.meta.ini}
\alias{do.plot.tfr.pardensity}
\alias{do.plot.tfr.partraces}
\alias{do.read.un.file}
\alias{do.write.projection.summary}
\alias{do.write.values.into.file}
\alias{extract.plot.args}
\alias{filter.traces}
\alias{find.tau.and.DLcountries}
\alias{get.all.parameter.names} 
\alias{get.all.parameter.names.cs}
\alias{get.all.parameter.names.extended} 
\alias{get.ar1.parameters}
\alias{get.ar1.countries}
\alias{get.ar1.countries.index}
\alias{get.ar1.data}
\alias{get.burned.tfr.traces}
\alias{get.countries.index}
\alias{get.data.for.worldmap}
\alias{get.data.for.worldmap.bayesTFR.prediction}
\alias{get.data.imputed}
\alias{get.data.for.country.imputed}
\alias{get.data.matrix}
\alias{get.friendly.variant.names}
\alias{get.full.par.names}
\alias{get.full.par.names.cs}
\alias{get.half.child.variant}
\alias{get.lambda}
\alias{get.mcmc.list}
\alias{get.mcmc.list.bayesTFR.mcmc}
\alias{get.mcmc.list.bayesTFR.mcmc.set}
\alias{get.mcmc.list.bayesTFR.prediction}
\alias{get.mcmc.list.list}
\alias{get.mcmc.meta}
\alias{get.median.from.prediction}
\alias{get.meta.only}
\alias{get.newTFRcolnames}
\alias{get.nr.countries}
\alias{get.nrest.countries}
\alias{get.observed.time.matrix.and.regions}
\alias{get.other.parameter.names} 
\alias{get.other.parameter.names.cs}
\alias{get.prediction.periods}
\alias{get.prediction.years}
\alias{get.projection.summary.header}
\alias{get.TFRmatrix.and.regions}
\alias{get.tfr.reconstructed}
\alias{get.tfr.shift}
\alias{get.tfr.shift.estimation}
\alias{get.thinned.burnin}
\alias{get.thinning.index}
\alias{get.totrans.parameter.names} 
\alias{get.totrans.parameter.names.cs}
\alias{get.traj.ascii.header}
\alias{get.traj.quantiles}
\alias{get.trans.parameter.names}
\alias{get.trans.parameter.names.cs}
\alias{get.trajectories}
\alias{get.UN.variant.names}
\alias{get_eps_T}
\alias{get_eps_T_all}
\alias{ini.tfr.with.DL}
\alias{init.nodes}
\alias{init.nodes.cont}
\alias{load.tfr.parameter.traces}
\alias{load.tfr.parameter.traces.all}
\alias{load.tfr.parameter.traces.cs}
\alias{log_cond_U}
\alias{log_cond_abf_sd}
\alias{log_cond_const_sd}
\alias{log_cond_d_trans}
\alias{log_cond_Triangle_c4_trans}
\alias{log_cond_sigma0}
\alias{log_cond_sigma0.tau}
\alias{log_like_gammas}
\alias{make.tfr.prediction}
\alias{mcmc.continue.chain}
\alias{mcmc.ini}
\alias{mcmc.ini.extra}
\alias{mcmc.meta.ini}
\alias{mcmc.meta.ini.extra}
\alias{mcmc.run.chain}
\alias{mcmc.run.chain.extra}
\alias{mcmc.update.U}
\alias{mcmc.update.abS}
\alias{mcmc.update.abSsigma0const}
\alias{mcmc.update.d}
\alias{mcmc.update.Triangle_c4}
\alias{mcmc.update.gamma}
\alias{mcmc.update.sigma0const}
\alias{mcmc.update.sigma0.tau}
\alias{no.traces.loaded}
\alias{overwrite.colnames}
\alias{par.names.for.worldmap}
\alias{process.not.converged.parameters}
\alias{read.tfr.file}
\alias{read.UNlocations}
\alias{read.UNtfr}
\alias{remove.tfr.traces}
\alias{set_wpp_regions}
\alias{set.wpp.extra}
\alias{stop.if.country.not.DL}
\alias{store.bayesTFR.meta.object}
\alias{store.bayesTFR.object}
\alias{store.mcmc}
\alias{store.traj.ascii}
\alias{thinindep}
\alias{tfr.get.dlcurves}
\alias{tfr.info}
\alias{tfr.mcmc.sampling}
\alias{tfr.mcmc.sampling.extra}
\alias{tfr.set.identical}
\alias{tfr.identical}
\alias{write.table.into.file.cdep}
\alias{write.table.into.file.cindep}
\alias{write.values.into.file.cdep}
\alias{write.values.into.file.cindep}

\alias{iso3166}
\alias{proposal_cov_gammas_cii}
\alias{correlation_predictors}
	
\title{
Internal Functions and datasets of bayesTFR 
}

\description{
These functions and datasets are not to be used directly by the user.
}

\keyword{internal}